# docs/swagger.yaml
basePath: /api
definitions:
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.TenantCreateRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  service.TenantUpdateRequest:
    properties:
      active:
        type: boolean
      monthly_request_limit:
        type: integer
      monthly_token_limit:
        type: integer
      name:
        type: string
    type: object
host: localhost:7847
info:
  contact: {}
  description: LLM 文档处理与知识图谱 API 中间件
  title: Nexus API
  version: "1.0"
paths:
...
  /admin/stats:
    get:
      description: 获取总览统计数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 仪表盘统计
      tags:
      - 统计
  /admin/stats/errors:
    get:
      description: 获取错误统计和分析数据
      parameters:
      - default: 7
        description: 统计天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 错误分析
      tags:
      - 统计
  /admin/stats/llm-performance:
    get:
      description: 获取 LLM 调用性能数据
      parameters:
      - default: 7
        description: 统计天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: LLM 性能统计
      tags:
      - 统计
  /admin/stats/pipeline-performance:
    get:
      description: 获取 Pipeline 处理性能数据
      parameters:
      - default: 7
        description: 统计天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: Pipeline 性能统计
      tags:
      - 统计
  /admin/tenants:
    get:
      description: 获取所有租户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取租户列表
      tags:
      - 租户
    post:
      consumes:
      - application/json
      description: 创建新租户
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.TenantCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建租户
      tags:
      - 租户
  /admin/tenants/{id}:
    delete:
      description: 删除指定租户
      parameters:
      - description: 租户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除租户
      tags:
      - 租户
    put:
      consumes:
      - application/json
      description: 更新租户名称、状态、配额等
      parameters:
      - description: 租户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.TenantUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新租户
      tags:
      - 租户
  /admin/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到存储服务
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 存储路径
        in: formData
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件
  /admin/webhooks:
    get:
      description: 分页获取 Webhook 投递日志
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取 Webhook 日志
      tags:
      - Webhook
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"